# Store number of labels as a field