--- a/src/array.jl
+++ b/src/array.jl
@@ -1,5 +1,6 @@
 """
-    OneHotArray{T, L, N, M, I} <: AbstractArray{Bool, M}
+    OneHotArray{T, N, M, I} <: AbstractArray{Bool, M}
+    OneHotArray(indices, L)
 
 A one-hot `M`-dimensional array with `L` labels (i.e. `size(A, 1) == L` and `sum(A, dims=1) == 1`)
 stored as a compact `N == M-1`-dimensional array of indices.
@@ -7,59 +8,70 @@ stored as a compact `N == M-1`-dimensional array of indices.
 Typically constructed by [`onehot`](@ref) and [`onehotbatch`](@ref).
 Parameter `I` is the type of the underlying storage, and `T` its eltype.
 """
-struct OneHotArray{T<:Integer, L, N, var"N+1", I<:Union{T, AbstractArray{T, N}}} <: AbstractArray{Bool, var"N+1"}
+struct OneHotArray{T<:Integer, N, var"N+1", I<:Union{T, AbstractArray{T, N}}} <: AbstractArray{Bool, var"N+1"}
   indices::I
+  nlabels::Int
 end
-OneHotArray{T, L, N, I}(indices) where {T, L, N, I} = OneHotArray{T, L, N, N+1, I}(indices)
-OneHotArray(indices::T, L::Integer) where {T<:Integer} = OneHotArray{T, L, 0, 1, T}(indices)
-OneHotArray(indices::I, L::Integer) where {T, N, I<:AbstractArray{T, N}} = OneHotArray{T, L, N, N+1, I}(indices)
+OneHotArray{T, N, I}(indices, L::Int) where {T, N, I} = OneHotArray{T, N, N+1, I}(indices, L)
+OneHotArray(indices::T, L::Int) where {T<:Integer} = OneHotArray{T, 0, 1, T}(indices, L)
+OneHotArray(indices::I, L::Int) where {T, N, I<:AbstractArray{T, N}} = OneHotArray{T, N, N+1, I}(indices, L)
 
 _indices(x::OneHotArray) = x.indices
-_indices(x::Base.ReshapedArray{<: Any, <: Any, <: OneHotArray}) =
+_indices(x::Base.ReshapedArray{<:Any, <:Any, <:OneHotArray}) =
   reshape(parent(x).indices, x.dims[2:end])
 
 """
-    OneHotVector{T, L} = OneHotArray{T, L, 0, 1, T}
+    OneHotVector{T} = OneHotArray{T, 0, 1, T}
+    OneHotVector(indices, L)
 
 A one-hot vector with `L` labels (i.e. `length(A) == L` and `count(A) == 1`) typically constructed by [`onehot`](@ref).
 Stored efficiently as a single index of type `T`, usually `UInt32`.
 """
-const OneHotVector{T, L} = OneHotArray{T, L, 0, 1, T}
+const OneHotVector{T} = OneHotArray{T, 0, 1, T}
+OneHotVector(idx, L) = OneHotArray(idx, L)
 
 """
-    OneHotMatrix{T, L, I} = OneHotArray{T, L, 1, 2, I}
+    OneHotMatrix{T, I} = OneHotArray{T, 1, 2, I}
+    OneHotMatrix(indices, L)
 
 A one-hot matrix (with `L` labels) typically constructed using [`onehotbatch`](@ref).
 Stored efficiently as a vector of indices with type `I` and eltype `T`.
 """
-const OneHotMatrix{T, L, I} = OneHotArray{T, L, 1, 2, I}
-
-OneHotVector(idx, L) = OneHotArray(idx, L)
+const OneHotMatrix{T, I} = OneHotArray{T, 1, 2, I}
 OneHotMatrix(indices, L) = OneHotArray(indices, L)
 
 # use this type so reshaped arrays hit fast paths
 # e.g. argmax
-const OneHotLike{T, L, N, var"N+1", I} =
-  Union{OneHotArray{T, L, N, var"N+1", I},
-        Base.ReshapedArray{Bool, var"N+1", <:OneHotArray{T, L, <:Any, <:Any, I}}}
+const OneHotLike{T, N, var"N+1", I} =
+  Union{OneHotArray{T, N, var"N+1", I},
+        Base.ReshapedArray{Bool, var"N+1", <:OneHotArray{T, <:Any, <:Any, I}}}
 
 _isonehot(x::OneHotArray) = true
-_isonehot(x::Base.ReshapedArray{<:Any, <:Any, <:OneHotArray{<:Any, L}}) where L = (size(x, 1) == L)
+_isonehot(x::Base.ReshapedArray{<:Any, <:Any, <:OneHotArray}) = (size(x, 1) == parent(x).nlabels)
 
-Base.size(x::OneHotArray{<:Any, L}) where L = (Int(L), size(x.indices)...)
+_check_nlabels(x::OneHotLike) = size(x, 1)
+function _check_nlabels(x::OneHotLike, xs::OneHotLike...)
+  L = size(x, 1)
+  all(size.(xs, 1) .== L) ||
+    throw(DimensionMismatch("The number of labels are not the same for all one-hot arrays."))
 
-function Base.getindex(x::OneHotArray{<:Any, <:Any, N}, i::Integer, I::Vararg{Any, N}) where N
+  return L
+end
+
+Base.size(x::OneHotArray) = (x.nlabels, size(x.indices)...)
+
+function Base.getindex(x::OneHotArray{<:Any, N}, i::Integer, I::Vararg{Any, N}) where N
   @boundscheck checkbounds(x, i, I...)
   return x.indices[I...] .== i
 end
 
-function Base.getindex(x::OneHotArray{<:Any, L}, ::Colon, I...) where L
+function Base.getindex(x::OneHotArray, ::Colon, I::Union{Colon, Integer}...)
   @boundscheck checkbounds(x, :, I...)
-  return OneHotArray(x.indices[I...], L)
+  return OneHotArray(x.indices[I...], x.nlabels)
 end
 
 Base.getindex(x::OneHotArray, ::Colon) = BitVector(reshape(x, :))
-Base.getindex(x::OneHotArray{<:Any, <:Any, N}, ::Colon, ::Vararg{Colon, N}) where N = x
+Base.getindex(x::OneHotArray{<:Any, N}, ::Colon, ::Vararg{Colon, N}) where N = x
 
 function Base.showarg(io::IO, x::OneHotArray, toplevel)
   print(io, ndims(x) == 1 ? "OneHotVector(" : ndims(x) == 2 ? "OneHotMatrix(" : "OneHotArray(")
@@ -77,17 +89,18 @@ end
 # copy CuArray versions back before trying to print them:
 for fun in (:show, :print_array)  # print_array is used by 3-arg show
   @eval begin
-    Base.$fun(io::IO, X::OneHotLike{T, L, N, var"N+1", <:AbstractGPUArray}) where {T, L, N, var"N+1"} = 
+    Base.$fun(io::IO, X::OneHotLike{T, N, var"N+1", <:AbstractGPUArray}) where {T, N, var"N+1"} = 
       Base.$fun(io, adapt(Array, X))
-    Base.$fun(io::IO, X::LinearAlgebra.AdjOrTrans{Bool, <:OneHotLike{T, L, N, <:Any, <:AbstractGPUArray}}) where {T, L, N} = 
+    Base.$fun(io::IO, X::LinearAlgebra.AdjOrTrans{Bool, <:OneHotLike{T, N, <:Any, <:AbstractGPUArray}}) where {T, N} = 
       Base.$fun(io, adapt(Array, X))
   end
 end
 
-_onehot_bool_type(::OneHotLike{<:Any, <:Any, <:Any, var"N+1", <:Union{Integer, AbstractArray}}) where {var"N+1"} = Array{Bool, var"N+1"}
-_onehot_bool_type(::OneHotLike{<:Any, <:Any, <:Any, var"N+1", <:AbstractGPUArray}) where {var"N+1"} = AbstractGPUArray{Bool, var"N+1"}
+_onehot_bool_type(::OneHotLike{<:Any, <:Any, var"N+1", <:Union{Integer, AbstractArray}}) where {var"N+1"} = Array{Bool, var"N+1"}
+_onehot_bool_type(::OneHotLike{<:Any, <:Any, var"N+1", <:AbstractGPUArray}) where {var"N+1"} = AbstractGPUArray{Bool, var"N+1"}
 
-function Base.cat(x::OneHotLike{<:Any, L}, xs::OneHotLike{<:Any, L}...; dims::Int) where L
+function Base.cat(x::OneHotLike, xs::OneHotLike...; dims::Int)
+  L = _check_nlabels(x, xs...)
   if isone(dims) || any(x -> !_isonehot(x), (x, xs...))
     return cat(map(x -> convert(_onehot_bool_type(x), x), (x, xs...))...; dims = dims)
   else
@@ -99,16 +112,16 @@ Base.hcat(x::OneHotLike, xs::OneHotLike...) = cat(x, xs...; dims = 2)
 Base.vcat(x::OneHotLike, xs::OneHotLike...) = cat(x, xs...; dims = 1)
 
 # optimized concatenation for matrices and vectors of same parameters
-Base.hcat(x::T, xs::T...) where {L, T <: OneHotLike{<:Any, L, <:Any, 2}} =
-  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), L)
-Base.hcat(x::T, xs::T...) where {L, T <: OneHotLike{<:Any, L, <:Any, 1}} =
-  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), L)
+Base.hcat(x::OneHotMatrix, xs::OneHotMatrix...) =
+  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _check_nlabels(x, xs...))
+Base.hcat(x::T, xs::T...) where {T <: OneHotLike{<:Any, <:Any, 1}} =
+  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _check_nlabels(x, xs...))
 
-MLUtils.batch(xs::AbstractArray{<:OneHotVector{<:Any, L}}) where L = OneHotMatrix(_indices.(xs), L)
+MLUtils.batch(xs::AbstractArray{<:OneHotVector}) = OneHotMatrix(_indices.(xs), _check_nlabels(xs...))
 
-Adapt.adapt_structure(T, x::OneHotArray{<:Any, L}) where L = OneHotArray(adapt(T, _indices(x)), L)
+Adapt.adapt_structure(T, x::OneHotArray) = OneHotArray(adapt(T, _indices(x)), x.nlabels)
 
-function Base.BroadcastStyle(::Type{<:OneHotArray{<: Any, <: Any, <: Any, var"N+1", T}}) where {var"N+1", T <: AbstractGPUArray}
+function Base.BroadcastStyle(::Type{<:OneHotArray{<:Any, <:Any, var"N+1", T}}) where {var"N+1", T <: AbstractGPUArray}
   # We want CuArrayStyle{N+1}(). There's an AbstractGPUArrayStyle but it doesn't do what we need. 
   S = Base.BroadcastStyle(T)
   # S has dim N not N+1. The following hack to fix it relies on the arraystyle having N as its first type parameter, which