--- a/src/array.jl
+++ b/src/array.jl
@@ -49,10 +49,12 @@ const OneHotLike{T, N, var"N+1", I} =
 _isonehot(x::OneHotArray) = true
 _isonehot(x::Base.ReshapedArray{<:Any, <:Any, <:OneHotArray}) = (size(x, 1) == parent(x).nlabels)
 
-_check_nlabels(x::OneHotLike) = size(x, 1)
-function _check_nlabels(x::OneHotLike, xs::OneHotLike...)
+_check_nlabels(L, xs::OneHotLike...) = all(size.(xs, 1) .== L)
+
+_nlabels(x::OneHotArray) = size(x, 1)
+function _nlabels(x::OneHotLike, xs::OneHotLike...)
   L = size(x, 1)
-  all(size.(xs, 1) .== L) ||
+  _check_nlabels(L, xs...) ||
     throw(DimensionMismatch("The number of labels are not the same for all one-hot arrays."))
 
   return L
@@ -61,13 +63,20 @@ end
 Base.size(x::OneHotArray) = (x.nlabels, size(x.indices)...)
 
 function Base.getindex(x::OneHotArray{<:Any, N}, i::Int, I::Vararg{Int, N}) where N
-  @boundscheck 1 <= i <= x.nlabels
+  @boundscheck (1 <= i <= x.nlabels) || throw(BoundsError(x, (i, I...)))
+  return x.indices[I...] .== i
+end
+# the method above is faster on the CPU but will scalar index on the GPU
+# so we define the method below to pass the extra indices directly to GPU array
+function Base.getindex(x::OneHotArray{<:Any, N, <:Any, <:AbstractGPUArray},
+                       i::Int, 
+                       I::Vararg{Any, N}) where N
+  @boundscheck (1 <= i <= x.nlabels) || throw(BoundsError(x, (i, I...)))
   return x.indices[I...] .== i
 end
 function Base.getindex(x::OneHotArray{<:Any, N}, ::Colon, I::Vararg{Any, N}) where N
   return OneHotArray(x.indices[I...], x.nlabels)
 end
-
 Base.getindex(x::OneHotArray, ::Colon) = BitVector(reshape(x, :))
 Base.getindex(x::OneHotArray{<:Any, N}, ::Colon, ::Vararg{Colon, N}) where N = x
 
@@ -97,25 +106,74 @@ end
 _onehot_bool_type(::OneHotLike{<:Any, <:Any, var"N+1", <:Union{Integer, AbstractArray}}) where {var"N+1"} = Array{Bool, var"N+1"}
 _onehot_bool_type(::OneHotLike{<:Any, <:Any, var"N+1", <:AbstractGPUArray}) where {var"N+1"} = AbstractGPUArray{Bool, var"N+1"}
 
-function Base.cat(x::OneHotLike, xs::OneHotLike...; dims::Int)
-  if isone(dims) || any(x -> !_isonehot(x), (x, xs...))
-    return cat(map(x -> convert(_onehot_bool_type(x), x), (x, xs...))...; dims = dims)
+_chk_cat_dim(d, x, y) = (x == y) ? x :
+  throw(DimensionMismatch("mismatch in dimension $d (expected $x got $y)"))
+
+_cat_size(iscatdims, dims::NTuple{N}) where N = dims
+function _cat_size(iscatdims, dims::NTuple{N}, x) where N
+  dims = ntuple(Val(N)) do i
+    d = dims[i]
+    sx = size(x, i)
+    iscatdims[i] ? d + sx : _chk_cat_dim(i, d, sx)
+  end
+
+  return dims
+end
+function _cat_size(iscatdims, dims::NTuple{N}, x, tail...) where N
+  dims = _cat_size(iscatdims, dims, x)
+  return _cat_size(iscatdims, dims, tail...)
+end
+
+_cat_similar(::Integer, sz) = zeros(Bool, sz)
+_cat_similar(x, sz) = similar(x, Bool, sz)
+
+_notall_onehot(x::OneHotArray, xs::OneHotArray...) = false
+_notall_onehot(x::OneHotLike, xs::OneHotLike...) = any(x -> !_isonehot(x), (x, xs...))
+
+function _cat_fallback(x::OneHotLike{<:Any, <:Any, N}, xs::OneHotLike...; dims::Int) where N
+  # this is adapted from Base
+  # https://github.com/JuliaLang/julia/blob/5544a0fab7648cfa61fe79cd557a7504a92ec1b5/base/abstractarray.jl#L1730-L1738
+  # invoking the less specialized version of cat results in even worse performance
+  # we really only want to be in this method if dims != 1 and all the labels match
+  # without the number of labels in the type, we can't do this at compile time
+  # so we reimplement some stuff from the less specialized cat
+  catdims = ntuple(in(dims), Val(N))
+  catsize = _cat_size(catdims, size(x), xs...)
+  y = _cat_similar(_indices(x), catsize)
+  offset = ones(Int, N)
+  for xi in (x, xs...)
+    nelements = size(xi)
+    I = ntuple(N) do i
+      catdims[i] ? (offset[i]:(offset[i] + nelements[i] - 1)) : Colon()
+    end
+    y[I...] = xi
+    offset .+= nelements
+  end
+
+  return y
+end
+
+function Base.cat(x::OneHotLike{<:Any, <:Any, N}, xs::OneHotLike...; dims::Int) where N
+  if isone(dims) || _notall_onehot(x, xs...)
+    return _cat_fallback(x, xs...; dims = dims)
   else
-    L = _check_nlabels(x, xs...)
+    L = _nlabels(x, xs...)
+
     return OneHotArray(cat(_indices(x), _indices.(xs)...; dims = dims - 1), L)
   end
 end
 
 Base.hcat(x::OneHotLike, xs::OneHotLike...) = cat(x, xs...; dims = 2)
-Base.vcat(x::OneHotLike, xs::OneHotLike...) = cat(x, xs...; dims = 1)
+Base.vcat(x::OneHotLike, xs::OneHotLike...) =
+  _cat_fallback(x, xs...; dims = 1)
 
 # optimized concatenation for matrices and vectors of same parameters
 Base.hcat(x::OneHotMatrix, xs::OneHotMatrix...) =
-  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _check_nlabels(x, xs...))
-Base.hcat(x::T, xs::T...) where {T <: OneHotLike{<:Any, <:Any, 1}} =
-  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _check_nlabels(x, xs...))
+  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _nlabels(x, xs...))
+Base.hcat(x::OneHotVector, xs::OneHotVector...) =
+  OneHotMatrix(reduce(vcat, _indices.(xs); init = _indices(x)), _nlabels(x, xs...))
 
-MLUtils.batch(xs::AbstractArray{<:OneHotVector}) = OneHotMatrix(_indices.(xs), _check_nlabels(xs...))
+MLUtils.batch(xs::AbstractArray{<:OneHotVector}) = OneHotMatrix(_indices.(xs), _nlabels(xs...))
 
 Adapt.adapt_structure(T, x::OneHotArray) = OneHotArray(adapt(T, _indices(x)), x.nlabels)