As suggested on Slack, storing the number of labels as a field instead of in the type results in type stable code. The [original PR](https://github.com/FluxML/Flux.jl/pull/1447) that I based my overhaul of one-hot arrays on mentions memory consumption as the main reason for keeping the label in the type. While it is certainly significantly less memory than the _old, old_ Flux implementation, compared to this PR, the overhead is constant (8 bytes).

I benchmarked to see if storing in the type resulted in downstream performance improvements (a reason to reject this PR). The results are mixed, but I think any measurable performance loss is due to a slower path being taken for indexing like `ov[1:3]`. I haven't figured how to fix this, but I think it is fixable. I don't think the performance loss is fundamental to where the number of label info is stored.

<details>
<summary> Benchmarking results and code </summary>

| Test target | Code | `@btime` for main | `@btime` for PR |
|:------------|:-----|:------------------|:----------------|
| `getindex` | `ov[1]` | 2.125 ns (0 allocations: 0 bytes) | 3.041 ns (0 allocations: 0 bytes) |
| `getindex` | `ov[1:3]` | 18.579 ns (1 allocation: 64 bytes) | 21.523 ns (1 allocation: 64 bytes) |
| `getindex` | `om[1, 2]` | 2.083 ns (0 allocations: 0 bytes) | 3.000 ns (0 allocations: 0 bytes) |
| `getindex` | `om[1:2, 3]` | 20.979 ns (1 allocation: 64 bytes) | 21.648 ns (1 allocation: 64 bytes) |
| `getindex` | `oa[1, 2, 3]` | 2.416 ns (0 allocations: 0 bytes) | 3.000 ns (0 allocations: 0 bytes) |
| `getindex` | `oa[1, :, 3]` | 201.261 ns (6 allocations: 256 bytes) | 24.556 ns (1 allocation: 64 bytes) |
| `getindex` | `oa[:, :, :]` | 2.125 ns (0 allocations: 0 bytes) | 2.125 ns (0 allocations: 0 bytes) |
| `size` | `length(ov)` | 1.167 ns (0 allocations: 0 bytes) | 2.125 ns (0 allocations: 0 bytes) |
| `size` | `size(oa)` | 2.083 ns (0 allocations: 0 bytes) | 2.084 ns (0 allocations: 0 bytes) |
| concat | `vcat(ov, ov2)` | 31.816 ns (1 allocation: 80 bytes) | 83.678 ns (3 allocations: 208 bytes) |
| concat | `hcat(ov, ov)` | 243.719 ns (2 allocations: 96 bytes) | 17.994 ns (1 allocation: 80 bytes) |
| concat | `hcat(om, om)` | 258.048 ns (2 allocations: 160 bytes) | 31.061 ns (1 allocation: 144 bytes) |
| concat | `vcat(om, om2)` | 186.143 ns (1 allocation: 160 bytes) | 272.052 ns (3 allocations: 384 bytes) |
| concat | `cat(oa, oa, oa; dims = 2)` | 2.093 μs (51 allocations: 1.98 KiB) | 2.528 μs (64 allocations: 2.66 KiB) |
| broadcast | `2 .* om` | 42.886 ns (1 allocation: 496 bytes) | 53.245 ns (1 allocation: 496 bytes) |
| broadcast | `om .* om .+ ov` | 72.852 ns (1 allocation: 496 bytes) | 116.178 ns (1 allocation: 496 bytes) |
| `argmax` | `argmax(om; dims = 1)`  | 34.876 ns (3 allocations: 240 bytes) | 35.211 ns (3 allocations: 240 bytes) |
| `argmax` | `argmax(oa; dims = 2)`  | 3.391 μs (28 allocations: 3.23 KiB) | 3.828 μs (28 allocations: 3.12 KiB) |
| onehot | `onehot((rand(1:10)), (1:10))` | 744.877 ns (8 allocations: 368 bytes) | 3.000 ns (0 allocations: 0 bytes) |
| onehot | `onehotbatch((rand(1:10, 50)), (1:10))` | 628.188 ns (4 allocations: 512 bytes) | 441.919 ns (4 allocations: 528 bytes) |
| onecold | `onecold(ov)` | 2.778 μs (19 allocations: 1.69 KiB) | 2.602 μs (19 allocations: 1.58 KiB) |
| onecold | `onecold(om)` | 22.590 ns (1 allocation: 96 bytes) | 22.757 ns (1 allocation: 96 bytes) |
| onecold | `onecold(oa)` | 34.743 ns (1 allocation: 256 bytes) | 34.660 ns (1 allocation: 256 bytes) |
| matmul | `(rand(5, 10)) * om` | 55.752 ns (1 allocation: 256 bytes) | 55.923 ns (1 allocation: 256 bytes) |
| matmul | `(rand(5, 10)) * ov` | 3.026 μs (28 allocations: 1.92 KiB) | 2.847 μs (28 allocations: 1.81 KiB) |
| matmul | `(rand(5, 5)) * adjoint(om)` | 53.879 ns (1 allocation: 496 bytes) | 53.583 ns (1 allocation: 496 bytes) |

Created by running:
```julia
using OneHotArrays
using BenchmarkTools

ov = OneHotVector(rand(1:10), 10)
ov2 = OneHotVector(rand(1:11), 11)
om = OneHotMatrix(rand(1:10, 5), 10)
om2 = OneHotMatrix(rand(1:11, 5), 11)
oa = OneHotArray(rand(1:10, 5, 5), 10)

@info "getindex"
@btime $ov[1]
@btime $ov[1:3]
@btime $om[1, 2]
@btime $om[1:2, 3]
@btime $oa[1, 2, 3]
@btime $oa[1, :, 3]
@btime $oa[:, :, :]

@info "size"
@btime length($ov)
@btime size($oa)

@info "concat"
@btime vcat($ov, $ov2)
@btime hcat($ov, $ov)
@btime hcat($om, $om)
@btime vcat($om, $om2)
@btime cat($oa, $oa, $oa; dims = 2)

@info "broadcast"
@btime 2 .* $om
@btime $om .* $om .+ $ov

@info "argmax"
@btime argmax($om; dims = 1)
@btime argmax($oa; dims = 2)

@info "onehot"
@btime onehot($(rand(1:10)), $(1:10))
@btime onehotbatch($(rand(1:10, 50)), $(1:10))

@info "onecold"
@btime onecold($ov)
@btime onecold($om)
@btime onecold($oa)

@info "matmul"
@btime $(rand(5, 10)) * $om
@btime $(rand(5, 10)) * $ov
@btime $(rand(5, 5)) * adjoint($om)
;
```

</details>