@ToucheSir 2022-10-08
These look like a nice set of benchmarks for future use too! I wonder if we can automate or at least script them somehow.