@darsnack 2022-10-11
> These look like a nice set of benchmarks for future use too! I wonder if we can automate or at least script them somehow.

We could definitely add a regression test for these. It doesn't take long to run them, so we can do a GH PR bot.