@ToucheSir 2022-10-12
I've been looking into https://github.com/tkf/BenchmarkCI.jl for other packages. This one seems like a nice place to test it.